+++ src/main/java/org/aion4j/avm/idea/service/AvmService.java
@@ -2,7 +2,8 @@ package org.aion4j.avm.idea.service;
 
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
-import org.jetbrains.idea.maven.project.MavenProject;
+
+import java.util.List;
 
 public interface AvmService {
 
@@ -15,6 +16,8 @@ public interface AvmService {
     public boolean isAvmProject();
 
     public boolean isClassAllowed(Project project, String clazz);
+
+    public List<MethodDescriptor> getAllowedMethodsForClass(Project project, String clazz, String methodName);
 
     public boolean isUnderTestSource(VirtualFile file);
 }