+++ src/main/java/org/aion4j/avm/idea/service/AvmServiceImpl.java
@@ -1,17 +1,27 @@
 package org.aion4j.avm.idea.service;
 
+import com.intellij.execution.ExecutionException;
+import com.intellij.execution.configurations.GeneralCommandLine;
+import com.intellij.execution.process.OSProcessHandler;
+import com.intellij.execution.process.ProcessAdapter;
+import com.intellij.execution.process.ProcessEvent;
+import com.intellij.execution.process.ProcessHandler;
+import com.intellij.ide.plugins.PluginManager;
 import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.extensions.PluginId;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.roots.ProjectRootManager;
+import com.intellij.openapi.util.Key;
 import com.intellij.openapi.vfs.VirtualFile;
-import com.intellij.psi.PsiClassType;
-import com.intellij.psi.PsiType;
-import com.intellij.psi.search.GlobalSearchScope;
-import org.aion4j.avm.idea.AvmDetails;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.idea.maven.model.MavenPlugin;
 import org.jetbrains.idea.maven.project.MavenProject;
 import org.jetbrains.idea.maven.project.MavenProjectsManager;
 
+import java.io.*;
+import java.nio.charset.Charset;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 public class AvmServiceImpl implements AvmService {
@@ -20,17 +30,23 @@ public class AvmServiceImpl implements AvmService {
     private static final Logger log = Logger.getInstance(AvmServiceImpl.class);
 
     private boolean isInitialized;
+    private boolean isMavenProject;
     private boolean isAvmProject;
 
     private List<String> testFolders;
     private List<String> sourceFolders;
 
-    public AvmServiceImpl() {
+    private boolean isJCLClassInitializationDone = false;
+//    private Map<String, Map<String, List<MethodDescriptor>>> jclWhitelist;
 
+    private JCLWhitelistHolder whitelistHolder;
+
+    public AvmServiceImpl() {
+        this.whitelistHolder = new JCLWhitelistHolder();
     }
 
     @Override
-    public void init(Project project) {
+    public synchronized void init(Project project) {
 
         this.isInitialized = true;
 
@@ -44,13 +60,16 @@ public class AvmServiceImpl implements AvmService {
             MavenPlugin mavenPlugin = rootProjects.get(0).findPlugin("org.aion4j", AION4j_MAVEN_PLUGIN);
             if(mavenPlugin == null) {
                 debug(() -> log.debug(">>>>>> Not a avm project"));;
+                //attachProjectListener(mvnProjectManager);
                 return;
             } else {
                 debug(() -> log.debug(">>>>>> It's a avm project"));
                 isAvmProject = true;
             }
         } else {
-            System.out.println("<<<<<<<<< No root project found");
+            if(log.isDebugEnabled()) {
+                log.debug("<<<<<<<<< No root project found");
+            }
             return;
         }
 
@@ -109,6 +128,27 @@ public class AvmServiceImpl implements AvmService {
         }*/
     }
 
+//    private void attachProjectListener(MavenProjectsManager projectsManager) {
+//
+//        projectsManager.addManagerListener(new MavenProjectsManager.Listener() {
+//            @Override
+//            public void activated() {
+//
+//            }
+//
+//            @Override
+//            public void projectsScheduled() {
+//
+//            }
+//
+//            @Override
+//            public void importAndResolveScheduled() {
+//                System.out.println("Scheduled for resolver..........");
+//
+//            }
+//        });
+//    }
+
     @Override
     public boolean isInitialize() {
         return isInitialized;
@@ -129,17 +169,13 @@ public class AvmServiceImpl implements AvmService {
         if(!isAvmProject)
             return true;
 
-        Class clazz = null;
-        try {
-            clazz = Class.forName(clazzName);
-        } catch (ClassNotFoundException e) {
-            //debug(() -> log.debug(e.getMessage()));
-            return true; //don't need to highlight it as it will be done by the editor.
-        }
+        if(!isJCLClassInitializationDone)
+            getJCLListFromProjectLibrary(project);
 
-        AvmDetails.MethodDescriptor[] methodDescriptors = AvmDetails.getClassLibraryWhiteList().get(clazz);
+        if(whitelistHolder.size() == 0)
+            return true;
 
-        if(methodDescriptors !=  null)
+        if(whitelistHolder.isClassPresent(clazzName))
             return true;
         else {
             if(clazzName.startsWith("java.") || clazzName.startsWith("javax.") || clazzName.startsWith("org.xml"))
@@ -147,6 +183,19 @@ public class AvmServiceImpl implements AvmService {
             else
                 return true;
         }
+    }
+
+    @Override
+    @NotNull
+    public List<MethodDescriptor> getAllowedMethodsForClass(Project project, String clazz, String methodName) {
+        if(!isAvmProject)
+            return Collections.EMPTY_LIST;
+
+        if(!isJCLClassInitializationDone)
+            getJCLListFromProjectLibrary(project);
+
+        return whitelistHolder.getMethods(clazz, methodName);
+
     }
 
     @Override
@@ -165,6 +214,105 @@ public class AvmServiceImpl implements AvmService {
         }
 
         return false;
+    }
+
+    private void getJCLListFromProjectLibrary(Project project) {
+        String homePath = ProjectRootManager.getInstance(project).getProjectSdk().getHomePath();
+        File pluginPath = PluginManager.getPlugin(PluginId.getId("org.aion4j.avm")).getPath();
+
+
+        if(log.isDebugEnabled()) {
+            log.info("SDK Home path: >>>>>>>>>  " + homePath);
+            log.debug("Plugin Path >>>> " + pluginPath);
+        }
+
+        copyFile("/AvmDetailsGetter.class", pluginPath.getAbsolutePath());
+
+        ArrayList<String> cmds = new ArrayList<>();
+        cmds.add(homePath + File.separator + "bin/java");
+        cmds.add("-cp");
+        cmds.add("lib" + File.separatorChar + "*");
+        cmds.add("AvmDetailsGetter");
+
+        GeneralCommandLine generalCommandLine = new GeneralCommandLine(cmds);
+        generalCommandLine.setCharset(Charset.forName("UTF-8"));
+        generalCommandLine.setWorkDirectory(pluginPath);
+
+        ProcessHandler processHandler = null;
+        try {
+            processHandler = new OSProcessHandler(generalCommandLine);
+
+            processHandler.startNotify();
+            processHandler.addProcessListener(new ProcessAdapter() {
+
+                StringBuffer sb = new StringBuffer();
+                @Override
+                public void onTextAvailable(@NotNull ProcessEvent event, @NotNull Key outputType) {
+                    sb.append(event.getText());
+                }
+
+                @Override
+                public void processTerminated(@NotNull ProcessEvent event) {
+
+                    String jsonStr = sb.toString();
+
+                    whitelistHolder.init(project, jsonStr);
+                }
+            });
+        } catch (ExecutionException e) {
+            e.printStackTrace();
+        }
+
+        whitelistHolder.loadFromCache(project);
+
+        isJCLClassInitializationDone = true;
+    }
+
+    private String copyFile(String fileName, String destFolder) {
+
+        InputStream fileStream = null;
+        OutputStream out = null;
+        try {
+            // Read the file we're looking for
+            fileStream = this.getClass().getResourceAsStream(fileName);
+
+            if (fileStream == null) {
+                throw new RuntimeException(String.format("%s is not found in the plugin jar. ", fileName));
+            }
+
+            File targetFile
+                    = new File(destFolder, fileName);
+
+            out = new FileOutputStream(targetFile);
+
+            // Write the file to the temp file
+            byte[] buffer = new byte[1024];
+            int len = fileStream.read(buffer);
+            while (len != -1) {
+                out.write(buffer, 0, len);
+                len = fileStream.read(buffer);
+            }
+
+            // Return the path of this sweet new file
+            return targetFile.getAbsolutePath();
+
+        } catch (IOException e) {
+            throw new RuntimeException("Error copying " + fileName + "to " + destFolder, e);
+        } finally {
+            if (fileStream != null) {
+                try {
+                    fileStream.close();
+                } catch (IOException e) {
+                }
+            }
+
+            if (out != null) {
+                try {
+                    out.close();
+                } catch (IOException e) {
+                }
+            }
+        }
     }
 
     private void debug(Runnable doWhenDebug) {